Extract Rust highlighting tests to a test/output directory
