+++ src/highlight/tests/rust/attributes.rs
@@ -1,0 +1,5 @@
+#[derive(Parser, Serialize)]
+struct Opts {
+    #[clap(short, long, default_value = "partials/head.html")]
+    head: PathBuf,
+}
