+++ src/highlight/tests/rust/constants.rs
@@ -1,0 +1,3 @@
+const x: &str = "abc";
+const y: u32 = 123;
+const z: f32 = 1.0;
