+++ src/highlight/tests/rust/functions_and_macros.html
@@ -1,0 +1,5 @@
+<pre id="5e845049255f423296fd6f20449be0bc" class="rust"><code class="rust"><span class="keyword">fn</span> <span class="function">rust_attributes</span><span class="punctuation">(</span><span class="punctuation">)</span> <span class="punctuation">{</span>
+    <span class="function macro">assert</span><span class="function macro">!</span><span class="punctuation">(</span><span class="variable">cool_function</span><span class="punctuation">(</span><span class="punctuation">)</span><span class="punctuation">)</span><span class="punctuation">;</span>
+    <span class="function macro">println</span><span class="function macro">!</span><span class="punctuation">(</span><span class="punctuation">)</span>
+<span class="punctuation">}</span>
+</code></pre>
