+++ src/highlight/tests/rust/functions_and_macros.rs
@@ -1,0 +1,4 @@
+fn rust_attributes() {
+    assert!(cool_function());
+    println!()
+}
