+++ src/highlight/tests/rust/type_builtin.rs
@@ -1,0 +1,1 @@
+const x: &str = "abc";
